backend/requirements.txt
fastapi
uvicorn[standard]
pymongo[srv]
python-dotenv
bcrypt
email-validator